Move DevUI classes to appropriate package
Console was overloaded to include both DevUI and the ANSI console stuff
