+++ core/devmode-spi/src/main/java/io/quarkus/dev/devui/DevConsoleManager.java
@@ -1,4 +1,4 @@
-package io.quarkus.dev.console;
+package io.quarkus.dev.devui;
 
 import java.util.Collections;
 import java.util.Map;
