+++ extensions/grpc/runtime/src/main/java/io/quarkus/grpc/runtime/devmode/GrpcServices.java
@@ -15,7 +15,7 @@ import grpc.health.v1.HealthOuterClass.HealthCheckResponse.ServingStatus;
 import io.grpc.MethodDescriptor.MethodType;
 import io.grpc.ServerMethodDefinition;
 import io.quarkus.arc.Subclass;
-import io.quarkus.dev.console.DevConsoleManager;
+import io.quarkus.dev.devui.DevConsoleManager;
 import io.quarkus.grpc.runtime.GrpcServerRecorder;
 import io.quarkus.grpc.runtime.GrpcServerRecorder.GrpcServiceDefinition;
 import io.quarkus.grpc.runtime.config.GrpcConfiguration;
