+++ extensions/vertx-http/deployment/src/main/java/io/quarkus/vertx/http/deployment/devmode/console/DevConsoleHttpHandler.java
@@ -22,8 +22,8 @@ import io.netty.handler.codec.http.HttpResponse;
 import io.netty.handler.codec.http.HttpVersion;
 import io.netty.handler.codec.http.LastHttpContent;
 import io.netty.util.ReferenceCountUtil;
-import io.quarkus.dev.console.DevConsoleRequest;
-import io.quarkus.dev.console.DevConsoleResponse;
+import io.quarkus.dev.devui.DevConsoleRequest;
+import io.quarkus.dev.devui.DevConsoleResponse;
 import io.quarkus.netty.runtime.virtual.VirtualAddress;
 import io.quarkus.netty.runtime.virtual.VirtualClientConnection;
 import io.quarkus.netty.runtime.virtual.VirtualResponseHandler;
